backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
python-multipart
black
flake8
isort
pre-commit
importlib
